wip move bpmnModel except code out of JsonTestUtils for future reuse
